# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: MediFlow — Multi-tenant AI Healthcare Appointment Platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tenants:
    get:
      operationId: listTenants
      tags: [tenants]
      summary: List all tenants
      responses:
        "200":
          description: List of tenants
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Tenant"
    post:
      operationId: createTenant
      tags: [tenants]
      summary: Create a new tenant
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TenantInput"
      responses:
        "201":
          description: Created tenant
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tenant"
  /tenants/{id}:
    get:
      operationId: getTenant
      tags: [tenants]
      summary: Get tenant by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Tenant
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tenant"
        "404":
          description: Not found
    patch:
      operationId: updateTenant
      tags: [tenants]
      summary: Update tenant
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TenantUpdate"
      responses:
        "200":
          description: Updated tenant
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tenant"
  /doctors:
    get:
      operationId: listDoctors
      tags: [doctors]
      summary: List doctors
      parameters:
        - name: tenantId
          in: query
          schema:
            type: integer
        - name: specialty
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of doctors
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Doctor"
    post:
      operationId: createDoctor
      tags: [doctors]
      summary: Create a new doctor
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DoctorInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Doctor"
  /doctors/{id}:
    get:
      operationId: getDoctor
      tags: [doctors]
      summary: Get doctor by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Doctor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Doctor"
        "404":
          description: Not found
    patch:
      operationId: updateDoctor
      tags: [doctors]
      summary: Update doctor
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DoctorUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Doctor"
    delete:
      operationId: deleteDoctor
      tags: [doctors]
      summary: Delete doctor
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /doctors/{id}/slots/{date}:
    get:
      operationId: listDoctorSlots
      tags: [doctors]
      summary: Get available slots for a doctor on a date
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: date
          in: path
          required: true
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Available time slots
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TimeSlot"
  /patients:
    get:
      operationId: listPatients
      tags: [patients]
      summary: List patients
      parameters:
        - name: tenantId
          in: query
          schema:
            type: integer
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of patients
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Patient"
    post:
      operationId: createPatient
      tags: [patients]
      summary: Create a new patient
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatientInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Patient"
  /patients/{id}:
    get:
      operationId: getPatient
      tags: [patients]
      summary: Get patient by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Patient
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Patient"
        "404":
          description: Not found
    patch:
      operationId: updatePatient
      tags: [patients]
      summary: Update patient
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatientUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Patient"
    delete:
      operationId: deletePatient
      tags: [patients]
      summary: Delete patient
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /dashboard/stats:
    get:
      operationId: getDashboardStats
      tags: [dashboard]
      summary: Get dashboard statistics
      parameters:
        - name: tenantId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Dashboard stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardStats"
  /dashboard/activity:
    get:
      operationId: getDashboardActivity
      tags: [dashboard]
      summary: Get recent activity feed
      parameters:
        - name: tenantId
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: Activity items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ActivityItem"
...
components:
  schemas:
...
    Tenant:
      type: object
      required: [id, name, slug, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        address:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        logoUrl:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    TenantInput:
      type: object
      required: [name, slug]
      properties:
        name:
          type: string
        slug:
          type: string
        address:
          type: string
        phone:
          type: string
        email:
          type: string
        logoUrl:
          type: string
    TenantUpdate:
      type: object
      properties:
        name:
          type: string
        address:
          type: string
        phone:
          type: string
        email:
          type: string
        logoUrl:
          type: string

    # Doctor
    Doctor:
      type: object
      required: [id, tenantId, name, specialty, createdAt]
      properties:
        id:
          type: integer
        tenantId:
          type: integer
        name:
          type: string
        specialty:
          type: string
        bio:
          type: ["string", "null"]
        licenseNumber:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        availableDays:
          type: string
        startTime:
          type: string
        endTime:
          type: string
        slotDurationMinutes:
          type: integer
        createdAt:
          type: string
          format: date-time
    DoctorInput:
      type: object
      required: [tenantId, name, specialty]
      properties:
        tenantId:
          type: integer
        name:
          type: string
        specialty:
          type: string
        bio:
          type: string
        licenseNumber:
          type: string
        phone:
          type: string
        email:
          type: string
        availableDays:
          type: string
        startTime:
          type: string
        endTime:
          type: string
        slotDurationMinutes:
          type: integer
    DoctorUpdate:
      type: object
      properties:
        name:
          type: string
        specialty:
          type: string
        bio:
          type: string
        licenseNumber:
          type: string
        phone:
          type: string
        email:
          type: string
        availableDays:
          type: string
        startTime:
          type: string
        endTime:
          type: string
        slotDurationMinutes:
          type: integer
    TimeSlot:
      type: object
      required: [date, startTime, endTime, available]
      properties:
        date:
          type: string
          format: date
        startTime:
          type: string
        endTime:
          type: string
        available:
          type: boolean

    # Patient
    Patient:
      type: object
      required: [id, tenantId, name, createdAt]
      properties:
        id:
          type: integer
        tenantId:
          type: integer
        name:
          type: string
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        dateOfBirth:
          type: ["string", "null"]
        bloodType:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    PatientInput:
      type: object
      required: [tenantId, name]
      properties:
        tenantId:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        dateOfBirth:
          type: string
        bloodType:
          type: string
        address:
          type: string
        notes:
          type: string
    PatientUpdate:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        dateOfBirth:
          type: string
        bloodType:
          type: string
        address:
          type: string
        notes:
          type: string

    # Appointment
...
    DashboardStats:
      type: object
      required: [totalAppointments, scheduledAppointments, completedAppointments, cancelledAppointments, totalDoctors, totalPatients, todayAppointments, appointmentsBySpecialty]
      properties:
        totalAppointments:
          type: integer
        scheduledAppointments:
          type: integer
        completedAppointments:
          type: integer
        cancelledAppointments:
          type: integer
        totalDoctors:
          type: integer
        totalPatients:
          type: integer
        todayAppointments:
          type: integer
        appointmentsBySpecialty:
          type: array
          items:
            $ref: "#/components/schemas/SpecialtyCount"
    SpecialtyCount:
      type: object
      required: [specialty, count]
      properties:
        specialty:
          type: string
        count:
          type: integer
    ActivityItem:
      type: object
      required: [id, type, message, createdAt]
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [appointment_booked, appointment_cancelled, appointment_completed, patient_registered, doctor_added]
        message:
          type: string
        patientName:
          type: ["string", "null"]
        doctorName:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time

    # AI
...