# lib/api-zod/src/generated/types/patient.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * MediFlow — Multi-tenant AI Healthcare Appointment Platform
 * OpenAPI spec version: 0.1.0
 */

export interface Patient {
  id: number;
  tenantId: number;
  name: string;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  dateOfBirth?: string | null;
  /** @nullable */
  bloodType?: string | null;
  /** @nullable */
  address?: string | null;
  /** @nullable */
  notes?: string | null;
  createdAt: Date;
}