# lib/api-zod/src/generated/types/tenant.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * MediFlow — Multi-tenant AI Healthcare Appointment Platform
 * OpenAPI spec version: 0.1.0
 */

export interface Tenant {
  id: number;
  name: string;
  slug: string;
  /** @nullable */
  address?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  logoUrl?: string | null;
  createdAt: Date;
}